requirements.txt
google-genai
langchain
langchain-openai

notebook

